requirements.txt
PyNaCl==1.5.0
mkdocs==1.6.0
mkdocs-material==9.5.31

